+++ storage/nodes.go
@@ -25,7 +25,13 @@ type node struct {
 	delegationsPerEpochPerParty map[string]map[string]pb.Delegation
 	rewardScoresPerEpoch        map[string]pb.RewardScore
 	rankingPerEpoch             map[string]pb.RankingScore
+	existsPerEpoch              map[string]bool // tells us whether this node existed in the given epoch
+	lastChangeAdded             bool            // the last state change (added/removed) experienced by this node
 	minEpoch                    *uint64
+}
+
+func (n *node) exists(epoch string) bool {
+	return n.existsPerEpoch[epoch]
 }
 
 type keyRotation struct {
@@ -71,15 +77,48 @@ func (ns *Node) ReloadConf(cfg Config) {
 	ns.Config = cfg
 }
 
-func (ns *Node) AddNode(n pb.Node) {
+// AddEpoch when we enter a new epoch we need to update the exists map for each node
+func (ns *Node) AddEpoch(epoch string) {
 	ns.mut.Lock()
 	defer ns.mut.Unlock()
+	// for all nodes copy their existence flags from the last known state change
+	for _, n := range ns.nodes {
+
+		if _, ok := n.existsPerEpoch[epoch]; ok {
+			// if we already know just move on, ValidatorUpdate event may have come through before the epoch event
+			continue
+		}
+		n.existsPerEpoch[epoch] = n.lastChangeAdded
+	}
+}
+
+func (ns *Node) AddNode(n pb.Node, added bool, fromEpoch uint64) {
+	ns.mut.Lock()
+	defer ns.mut.Unlock()
+
+	ns.log.Info("adding node", logging.String("nodeid", n.Id), logging.Bool("added", added), logging.Uint64("from", fromEpoch))
+	epochSeq := strconv.FormatUint(fromEpoch, 10)
+	haveNode, ok := ns.nodes[n.GetId()]
+	if ok {
+		// node already exists in our store just update its existence flag
+		ns.log.Info("aleady existss just update", logging.String("nodeid", n.Id), logging.Bool("added", added), logging.Uint64("from", fromEpoch))
+		haveNode.existsPerEpoch[epochSeq] = added
+		haveNode.lastChangeAdded = added
+		return
+	}
+
+	if !added && !ok {
+		ns.log.Error("node has been removed despite never existing", logging.String("nodeID", n.GetId()))
+		return
+	}
 
 	nd := node{
 		n:                           n,
 		rewardScoresPerEpoch:        map[string]vega.RewardScore{},
 		rankingPerEpoch:             map[string]vega.RankingScore{},
 		delegationsPerEpochPerParty: map[string]map[string]pb.Delegation{},
+		existsPerEpoch:              map[string]bool{epochSeq: true},
+		lastChangeAdded:             true,
 		minEpoch:                    new(uint64),
 	}
 	*nd.minEpoch = math.MaxUint64
@@ -140,7 +179,7 @@ func (ns *Node) GetByID(id, epochID string) (*pb.Node, error) {
 		return nil, ErrNodeDoesNotExist
 	}
 
-	if _, ok := node.rankingPerEpoch[epochID]; !ok {
+	if !node.exists(epochID) {
 		return nil, ErrNodeDoesNotExistInThisEpoch
 	}
 
@@ -154,7 +193,7 @@ func (ns *Node) GetAll(epochID string) []*pb.Node {
 
 	nodes := make([]*pb.Node, 0, len(ns.nodes))
 	for _, n := range ns.nodes {
-		if _, ok := n.rankingPerEpoch[epochID]; !ok {
+		if !n.exists(epochID) {
 			// node was removed due to inactivity and so does not exist in this epoch
 			continue
 		}
@@ -184,7 +223,7 @@ func (ns *Node) GetTotalNodesNumber(epochID string) int {
 
 	count := 0
 	for _, n := range ns.nodes {
-		if _, ok := n.rankingPerEpoch[epochID]; !ok {
+		if !n.exists(epochID) {
 			continue
 		}
 		count += 1